requirements.txt
django-rest-framework
django-filter
mysqlclient~=2.0.3
pydantic